* master:
  Adding autopep8 as a pre-commit hook. (#7071)
  Refresh proxy protocol diagram (#7095)
  Skip docs builds if there are no changes in the doc dir and files it includes (#7088)
  Remove more deadcode (#7098)
  destroy threads after job done (#7083)
  Fix compilation error - missing '&' operator (#7093)
  Adds description for ssl_ticket_number in ssl_multicert docs (#7091)
  Ran clang-tidy over the source tree (#7077)
  Move the direct self loop check later to HttpSM::do_http_server_open just before connection upstream. (#7069)

+++ plugins/lua/ts_lua.c
@@ -129,7 +129,7 @@ create_lua_vms()
       ts_lua_max_state_count = TS_LUA_MAX_STATE_COUNT;
     } else {
       ts_lua_max_state_count = (int)mgmt_state;
-      TSDebug(TS_LUA_DEBUG_TAG, "[%s] found %s: [%d]", __FUNCTION__, ts_lua_mgmt_state_str, (int)ts_lua_max_state_count);
+      TSDebug(TS_LUA_DEBUG_TAG, "[%s] found %s: [%d]", __FUNCTION__, ts_lua_mgmt_state_str, ts_lua_max_state_count);
     }
 
     if (ts_lua_max_state_count < 1) {
